completed (except macros)

+++ exercises/23_conversions/using_as.rs
@@ -10,11 +10,11 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
+// le résultat de la fonction average était éffectué avec un f64 et un usize, on force le compilateur à traier le calcul complet en tant que f64
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
